Fix setting of correct output type if -a is given after -b argment.
Set the correct output type based on the file given by -a, -abam or -i.

Before this patch, the output type was determined on the first
file. If -b was specified before -a, the first -b file would
be used instead of the -a file.

This fixes the first problem of issue https://github.com/arq5x/bedtools2/issues/497:
  "Switching order of -a and -b paramter in bedtools intersect when -b file is BAM file, outputs BAM header."