Ignore CR characters at the ends of lines
Similarly to code in bedFile.cpp/bedGraphFile.h/tabFile.cpp, remove any
final CR character, so that we accept both Unix-style LF-terminated and
DOS-style CRLF-terminated lines.  Do this in BufferedStreamMgr::getLine()
and also in FileRecordTypeChecker::handleTextFormat() so that a line such
as "chr 100 200<CR><LF>" is recognized as BED format.