Handle the case of bedType==0 (i.e., an empty input file)
Avoid looping SIZE_MAX times when _bed->bedType is 0; we can't just
subtract 1 from an unsigned variable with value 0. Fixes #816.

This still prints an extra tab when bedType is 0 as the code that
prints the labels prefixes the first one with a tab, but as this
is a degenerate case of an empty file it's not worth changing.

(There are no other instances of "bedType - 1" at present.)