Improve handling of q==0.0 case in kt_fisher_exact()
When some of kt_fisher_exact()'s input arguments are very large,
hypergeo(...) is dominated by the subtracted lbinom() so returns
exp(-750 or less), which underflows, hence hypergeo() returns 0.0.

Revert the previous naive attempt at fixing this. Instead to avoid
this problem a special case is added for q == 0.0 which determines
the output values by comparing n11 to the location of the peak of
the hypergeometric distribution.